payment fixed

+++ contracts/payroll-vault/src/interface.rs
@@ -49,7 +49,7 @@ pub trait VaultTrait {
         name: String,
         payment_period: PaymentPeriod, // enum weekly monthly or anually
         salary: i128,
-        notice_period: i128, // how many payment periods before the employee can be fired
+        notice_period: u64, // how many payment periods before the employee can be fired
     ) -> Result<(), ContractError>;
 
     fn pay_employees(
